[Oztechan/CCC#2604] Move iOS WatchersView to WatchersRootView
